Add unified VM header
